requirements.txt
livekit-agents[openai,silero,turn-detector,deepgram]~=1.0
livekit-plugins-noise-cancellation~=0.2
python-dotenv
pypdf
langchain
chromadb # Vector storage
pytest
boto3
langchain-openai